clara_ai/backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2,<3
python-dotenv>=1,<2
requests>=2.31,<3
python-multipart>=0.0.9,<1
openai-whisper>=20231117
ffmpeg-python>=0.2,<1
python-docx>=1.1,<2
reportlab>=4.2,<5
